Add walkmesh customizablity to settings

+++ toolset/gui/widgets/color_edit.py
@@ -19,7 +19,7 @@ class ColorEdit(QWidget):
 
     def openColorDialog(self) -> None:
         initColor = Color.from_rgba_integer(self.ui.colorSpin.value())
-        initQColor = QColor(initColor.r*255, initColor.g*255, initColor.b*255, initColor.a*255)
+        initQColor = QColor(int(initColor.r*255), int(initColor.g*255), int(initColor.b*255), int(initColor.a*255))
 
         dialog = QColorDialog(QColor(initQColor.red(), initQColor.green(), initQColor.blue(), initQColor.alpha()))
         dialog.setOption(QColorDialog.ShowAlphaChannel, on=self.allowAlpha)
